% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment6.R
\docType{data}
\name{data}
\alias{data}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{id}{}
\item{variance}{}
\item{hint}{}
\item{sex}{}
\item{age}{}
\item{project}{}
\item{ranking}{}
\item{allocation}{}
\item{npv_amount}{}
\item{sample}{}
\item{max_npv}{}
\item{min_npv}{}
\item{max_min_difference}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{alignment6}
}
