% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment6.R
\docType{package}
\name{alignment6}
\alias{alignment6}
\alias{alignment6-package}
\title{Alignment 6 experiment}
\description{
Data and materials for the Alignment 6 experiment.
}
\details{
alignment6
A data package for the Alignment 6 experiment.

Use \code{data(package='alignment6')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
